Rendering/bloom (#154)
* Bloom added done, post process updgrade done, new window for post process wip

* Widget Post Process Added  / need work on multi post process

+++ Engine/include/Window.hpp
@@ -18,14 +18,14 @@ namespace BwatEngine
 
 		GLFWwindow* handler = nullptr;
 
-		float GetHeight() { return height; };
-		float GetWidth() { return width; };
+		int GetHeight() { return height; };
+		int GetWidth() { return width; };
 		bool IsWorking() { return !glfwWindowShouldClose(handler); };
 
 	private:
 
-		float width = 0.f;
-		float height = 0.f;
+		int width = 0.f;
+		int height = 0.f;
 
 	};
